Move visitor tests to separate file
