+++ tests/python/__init__.py
@@ -1,2 +1,3 @@
 from .test_builder import BuilderTests
 from .test_parser import ParserTests
+from .test_visitor import VisitorTests
